0.4.0
- changed all numbers to `usize`
- changed calc_aspect_res() to actually use the damn function i made
- add colored crate but didn't actually implement it yet, shut up i'm lazy

+++ src/args.rs
@@ -5,11 +5,11 @@ use argh::FromArgs;
 pub struct Args {
     /// the width to calculate with
     #[argh(option, short = 'w')]
-    pub width: Option<i32>,
+    pub width: Option<usize>,
 
     /// the height to calculate with
     #[argh(option, short = 'h')]
-    pub height: Option<i32>,
+    pub height: Option<usize>,
 
     /// get the size and aspect of an image
     #[argh(option, short = 'i')]
